+++ libs/ui/src/components/atoms/Dropdown/index.ts
@@ -1,0 +1,2 @@
+export * from './Dropdown';
+export * from './Dropdown.stories';
