--- libs/ui/src/components/atoms/dropdown/index.ts
@@ -1,2 +1,0 @@
-export * from './dropdown';
-export * from './dropdown.stories';
